ログアウト時のCookie削除とDBからのユーザーデータ削除機能を追加
