+++ apps/api/src/index.ts
@@ -1,9 +1,8 @@
 import { Hono } from "hono";
 import { cors } from "hono/cors";
-import { getCookie, setCookie } from "hono/cookie";
+import { getCookie } from "hono/cookie";
 import { drizzle } from "drizzle-orm/d1"; // DB実装を有効化
 import * as schema from "../drizzle/schema"; // DB実装を有効化
-import { eq } from "drizzle-orm";
 
 import { conversations } from "./routes/conversation";
 import { websocket } from "./routes/websocket"; // 既存のwebsocketルーター
@@ -59,149 +58,11 @@ app.use("*", async (c, next) => {
 
 // ルートパス ("/") の変更
 // アプリケーションのルートURLにアクセスした際に、GitHub認証を開始するためのHTMLを返します。
-app.get("/", (c) => {
-  return c.html(`
-    <html lang="ja">
-      <head>
-        <title>Prism GitHub認証</title>
-        <script src="https://cdn.tailwindcss.com"></script>
-        <style>
-          body { font-family: 'Inter', sans-serif; }
-        </style>
-      </head>
-      <body class="bg-gray-900 text-white min-h-screen flex items-center justify-center p-4">
-        <div class="bg-gray-800 shadow-lg p-8 max-w-md w-full text-center rounded-xl">
-          <h1 class="text-3xl font-bold mb-4 text-indigo-400">Prism GitHub認証テスト</h1>
-          <p class="text-lg mb-6 text-gray-300">GitHubアカウントとの連携を開始します。</p>
-          <a href="/github/oauth" class="bg-green-600 hover:bg-green-700 text-white font-bold py-2 px-4 rounded-lg transition-colors duration-300 transform hover:scale-105 inline-block">
-            GitHubで認証を開始する
-          </a>
-          <div class="mt-8 text-gray-400">
-            <p>DBテスト用エンドポイント:</p>
-            <p><code class="bg-gray-700 p-1 rounded">/db/test-user</code></p>
-          </div>
-        </div>
-      </body>
-    </html>
-  `);
-});
-
-// ログアウトでデータ削除
-app.post("/auth/logout", async (c) => {
-  const userId = c.get("userId");
-  // セッションCookie削除
-  setCookie(c, "prism_uid", "", {
-    httpOnly: true,
-    secure: true,
-    sameSite: "Lax",
-    path: "/",
-    maxAge: 0,
-  });
-  // 念のため一時Cookieも削除（存在していれば）
-  setCookie(c, "ext_redirect", "", { path: "/", maxAge: 0 });
-  setCookie(c, "github_oauth_state", "", {
-    httpOnly: true,
-    secure: true,
-    sameSite: "Lax",
-    path: "/",
-    maxAge: 0,
-  });
-
-  // DBからユーザーデータ削除（cascadeで関連テーブルも削除）
-  if (userId) {
-    const db = c.get("db");
-    if (db) {
-      try {
-        await db.delete(schema.users).where(eq(schema.users.id, userId));
-        console.log(`User ${userId} and all related data deleted`);
-      } catch (error) {
-        console.error("Failed to delete user data:", error);
-        // エラーでもCookie削除は成功として扱う
-      }
-    }
-  }
-
-  return c.json({ success: true });
-});
-
 // GitHub認証関連のルートを登録
 app.route("/github", githubRouter);
 
 app.route("/", conversations);
 
 app.route("/ws", websocket); // 既存のwebsocketルーターの登録
-
-// 新規追加: DBテスト用エンドポイント
-app.get("/db/test-user", async (c) => {
-  const db = c.get("db"); // ミドルウェアで設定されたDBインスタンスを取得
-
-  if (!db) {
-    return c.text(
-      "Database not initialized. Check your D1 binding in wrangler.jsonc or .env.",
-      500,
-    );
-  }
-
-  try {
-    // 1. users テーブルにデータを挿入 (または既存ユーザーを更新)
-    const testGithubId = -12345; // テスト用のGitHub ID（本番IDと衝突しない負の値を使用）
-    const testGithubUsername = "testuser_prism"; // テスト用のユーザー名
-
-    // upsert (挿入または更新) のロジック
-    // まず既存ユーザーを検索
-    const existingUsers = await db
-      .select()
-      .from(schema.users)
-      .where(eq(schema.users.githubId, testGithubId))
-      .limit(1);
-    let user: typeof schema.users.$inferSelect;
-
-    if (existingUsers.length > 0) {
-      // ユーザーが存在する場合、更新
-      const [updated] = await db
-        .update(schema.users)
-        .set({
-          githubUsername: testGithubUsername,
-          updatedAt: new Date().toISOString(),
-        })
-        .where(eq(schema.users.githubId, testGithubId))
-        .returning();
-      if (!updated) return c.json({ error: "Update failed" }, 500);
-      user = updated;
-      console.log("Existing user updated:", user);
-    } else {
-      // ユーザーが存在しない場合、新規挿入
-      const [created] = await db
-        .insert(schema.users)
-        .values({
-          githubId: testGithubId,
-          githubUsername: testGithubUsername,
-        })
-        .returning();
-      if (!created) return c.json({ error: "Insert failed" }, 500);
-      user = created;
-      console.log("New user created:", user);
-    }
-
-    // 2. users テーブルからすべてのデータを取得して確認
-    const allUsers = await db.select().from(schema.users);
-    console.log("All users in DB:", allUsers);
-
-    return c.json({
-      message: "DB test successful: User inserted/updated and retrieved.",
-      insertedOrUpdatedUser: user,
-      allUsersInDb: allUsers,
-    });
-  } catch (error) {
-    console.error("DB test error:", error);
-    return c.json(
-      {
-        error: "Failed to perform DB operations.",
-        details: error instanceof Error ? error.message : String(error),
-      },
-      500,
-    );
-  }
-});
 
 export default app;
